Modify the look of general headers that are still awaiting approval
