--- frontend/src/pages/ContentPage/PageCard.css
@@ -1,3 +1,0 @@
-i:hover {
-  cursor: pointer;
-}
